remove outdated benchmarks
I ran some numbers on the current build and we're a bit slower
today than we were when I last ran benchmarks.  We've grown
more state in the CellAttributes that we manage so the profiler
shows those as relatively hot spots.

Performance is now on par with iTerm2 (w/ Metal Renderer) on macos
honestly now a little slower than vte based terminals, but still faster
than xterm (everything is faster than xterm!)

Alacritty is impressively holding its own at the front of the pack,
even with scrollback.

There's room for improvement in wezterm, but it still feels fine,
so perf isn't currently top of my list.

Closes https://github.com/wez/wezterm/issues/20