Revert "add snapcraft build config"
This reverts commit 01b01b6547fff70c74b95f4dc2cfebbcd3f514f8.

My experience with even getting in the door with snaps was sub-par.
The submission/review process seems under specified and under staffed.

The tools for producing snaps were also rather broken in the case
that your PATH doesn't match a blessed configuration.  This was
hard to discover and expensive to iterate on.

My conclusion is that snaps are not for me.